+++ src/exec/launch_program.c
@@ -6,7 +6,7 @@
 /*   By: gpollast <gpollast@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/11 18:37:31 by erpascua          #+#    #+#             */
-/*   Updated: 2025/09/13 19:55:03 by gpollast         ###   ########.fr       */
+/*   Updated: 2025/09/18 15:51:56 by gpollast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,8 +45,6 @@ static int	repl(t_msh *msh, int tmp_fd)
 		parse(msh);
 		stack_destroy(msh->stack);
 		msh->stack = NULL;
-		// Reset data pointer after use
-		msh->data = NULL;
 		write(tmp_fd, msh->entry, ft_strlen(msh->entry));
 		write(tmp_fd, "\n", 1);
 		free(msh->entry);
