+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: gpollast <gpollast@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/11 10:57:36 by erpascua          #+#    #+#             */
-/*   Updated: 2025/09/12 11:16:43 by gpollast         ###   ########.fr       */
+/*   Updated: 2025/09/18 17:06:09 by gpollast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,31 +16,48 @@ int	g_exit_code = 0;
 
 int	env_dup(t_msh *msh, char **env)
 {
-	int	len;
+    int	len;
+    int	i;
+
+    len = 0;
+    while (env[len])
+        len++;
+    msh->env = malloc(sizeof(char *) * (len + 1));
+    if (!msh->env)
+        return (g_exit_code = 2, 0);
+    i = 0;
+    while (env[i])
+    {
+        msh->env[i] = ft_strdup(env[i]);
+        if (!msh->env[i])
+        {
+            while (--i >= 0)
+                free(msh->env[i]);
+            free(msh->env);
+            msh->env = NULL;
+            return (g_exit_code = 2, 0);
+        }
+        i++;
+    }
+    msh->env[i] = NULL;
+    return (1);
+}
+
+// void	free_msh(t_msh *msh)
+// {
+	
+// }
+
+void	free_msh_builtins(t_msh *msh)
+{
 	int	i;
 
-	len = 0;
-	while (env[len])
-		len++;
-	msh->env = malloc(sizeof(char *) * (len + 1));
-	if (!msh->env)
-	{
-		g_exit_code = 2;
-		return (0);
-	}
 	i = 0;
-	while (env[i])
+	while (i < NB_BUILTINS)
 	{
-		msh->env[i] = ft_strdup(env[i]);
-		if (!msh->env[i])
-		{
-			g_exit_code = 2;			
-			return (0);
-		}
+		free(msh->builtin_names[i]);
 		i++;
 	}
-	msh->env[i] = NULL;
-	return (1);
 }
 
 int	main(int ac, char **av, char **env)
@@ -55,6 +72,10 @@ int	main(int ac, char **av, char **env)
 	if (!env_dup(&msh, env))
 		return (g_exit_code);
 	launch_program(&msh);
-	// ft_free(&msh);
+	data_destroy(msh.data);
+	if (msh.env)
+		free_tab(msh.env);
+	free_msh_builtins(&msh);
+	// free_msh(&msh);
 	return (g_exit_code);
 }
